pyproject.toml
[tool.poetry]
name = "pbot2"
version = "0.1.0"
description = ""
authors = ["Chris Cummings <chris@chriscummings.net>"]

[tool.poetry.dependencies]
python = "^3.7"
discord = "^1.7.3"
pytest = "^7.1.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = ">=^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
